Removed dependency on "Options" + Fixes.
Dependency on the package "options" was causing the system to be
incompatible with web deployment, as it in turn depended on fs. To that
end, I've rewritten the code to combine configurations to factor it out.

Implementation of the .connectTo method has begun, although there is still
some way to go. There are additional fixes littered throughout the code, I
can't recall them all just now.